Hide TextRange. It has no public usage.
